% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stat.R
\name{calc_stat}
\alias{calc_stat}
\title{Statistics Calculation for Average Medicare Payments}
\usage{
calc_stat(stat = "mean")
}
\arguments{
\item{stat}{Mean, median, or standard deviation}
}
\value{
Mean, median, or standard deviation of average medicare payments
}
\description{
This a function that calculates statistics(mean, median, or standard deviation) over
all of the DRG codes for average medicare payments.
}
\examples{
calc_stat(stat = "mean")
calc_stat(stat = "median")
calc_stat(stat = "sd")
}
